refactor h12 and h1x
