+++ malariagen_data/ag3.py
@@ -14,10 +14,7 @@ MAJOR_VERSION_NUMBER = 3
 MAJOR_VERSION_PATH = "v3"
 CONFIG_PATH = "v3-config.json"
 GCS_URL = "gs://vo_agam_release/"
-H12_CALIBRATION_CACHE_NAME = "ag3_h12_calibration_v1"
-H12_GWSS_CACHE_NAME = "ag3_h12_gwss_v1"
 XPEHH_GWSS_CACHE_NAME = "ag3_xpehh_gwss_v1"
-H1X_GWSS_CACHE_NAME = "ag3_h1x_gwss_v1"
 IHS_GWSS_CACHE_NAME = "ag3_ihs_gwss_v1"
 VIRTUAL_CONTIGS = {
     "2RL": ("2R", "2L"),
@@ -131,10 +128,7 @@ class Ag3(AnophelesDataResource):
 
     """
 
-    _h12_calibration_cache_name = H12_CALIBRATION_CACHE_NAME
-    _h12_gwss_cache_name = H12_GWSS_CACHE_NAME
     _xpehh_gwss_cache_name = XPEHH_GWSS_CACHE_NAME
-    _h1x_gwss_cache_name = H1X_GWSS_CACHE_NAME
     _ihs_gwss_cache_name = IHS_GWSS_CACHE_NAME
 
     def __init__(
